Part-3/requirements.txt
flask
gunicorn
python-binance
tradingview_ta
datetime
pandas
numpy
pandas_ta
rq
